Ensure records are properly imported when they require a write() instead of create() and other records fail due to DomainError.
Previous code did not save all the records if a write operation was required
and some of the records failed. It seems that although core's save() method
tries to restablish record's _values when there's an exception it does not
always work. At least it failed when we imported stock.inventory.line and one
of them failed due to DomainError (trying to put an 'asset' product in the
inventory line).

I reported it into core (https://bugs.b2ck.com/issue8085) but at least this
patch works around the problem by caching values and reassigning them if
something goes wrong.